+++ changelog.d/wa-eceap.added.md
@@ -1,1 +1,1 @@
-Implement Washington's Early Childhood Education and Assistance Program (ECEAP), including the standard ECEAP eligibility for ages 3-4 and Early ECEAP eligibility for birth to age 3.
+Implement Washington's Early Childhood Education and Assistance Program (ECEAP), including the standard ECEAP eligibility for ages 3-5 and Birth to Three ECEAP eligibility for children below age 3.
