+++ policyengine_us/parameters/gov/states/wa/dcyf/eceap/birth_to_three_eceap/age_limit.yaml
@@ -1,0 +1,14 @@
+description: Washington limits Birth to Three ECEAP participants to children below this age.
+
+values:
+  2024-01-01: 3
+
+metadata:
+  unit: year
+  period: year
+  label: Washington Birth to Three ECEAP children age limit
+  reference:
+    - title: RCW 43.216.505 Eligible children
+      href: https://app.leg.wa.gov/RCW/default.aspx?cite=43.216.505
+    - title: WAC 110-425 Birth to Three ECEAP definition (birth to age 3)
+      href: https://app.leg.wa.gov/WAC/default.aspx?cite=110-425
