--- policyengine_us/parameters/gov/states/wa/dcyf/eceap/early_eceap/age_limit.yaml
@@ -1,14 +1,0 @@
-description: Washington limits Early Childhood Education and Assistance Program participants under the birth-to-three variant to children below this age.
-
-values:
-  2024-01-01: 3
-
-metadata:
-  unit: year
-  period: year
-  label: Washington Early ECEAP children age limit
-  reference:
-    - title: RCW 43.216.505 Eligible children
-      href: https://app.leg.wa.gov/RCW/default.aspx?cite=43.216.505
-    - title: WAC 110-425 Early ECEAP definition (birth to age 3)
-      href: https://app.leg.wa.gov/WAC/default.aspx?cite=110-425
