--- policyengine_us/parameters/gov/states/wa/dcyf/eceap/early_eceap/income/fpg_rate.yaml
@@ -1,12 +1,0 @@
-description: Washington limits income to this fraction of the federal poverty guideline when determining family financial need under the Early Childhood Education and Assistance Program for children under age three (Birth to Three ECEAP, pre-2025-10-01 regime).
-
-values:
-  2021-07-25: 1.3
-
-metadata:
-  unit: /1
-  period: year
-  label: Washington Early ECEAP income threshold as fraction of federal poverty guideline
-  reference:
-    - title: RCW 43.216.505 Eligible children
-      href: https://app.leg.wa.gov/RCW/default.aspx?cite=43.216.505
