--- policyengine_us/tests/policy/baseline/gov/states/wa/dcyf/eceap/wa_early_eceap_age_eligible.yaml
@@ -1,47 +1,0 @@
-- name: Age 0, below Early ECEAP age limit.
-  period: 2025
-  input:
-    age: 0
-    state_code: WA
-  output:
-    wa_early_eceap_age_eligible: true
-
-- name: Age 1, below Early ECEAP age limit.
-  period: 2025
-  input:
-    age: 1
-    state_code: WA
-  output:
-    wa_early_eceap_age_eligible: true
-
-- name: Age 2, below Early ECEAP age limit.
-  period: 2025
-  input:
-    age: 2
-    state_code: WA
-  output:
-    wa_early_eceap_age_eligible: true
-
-- name: Age 3, at Early ECEAP age limit (not eligible).
-  period: 2025
-  input:
-    age: 3
-    state_code: WA
-  output:
-    wa_early_eceap_age_eligible: false
-
-- name: Age 4, above Early ECEAP age limit.
-  period: 2025
-  input:
-    age: 4
-    state_code: WA
-  output:
-    wa_early_eceap_age_eligible: false
-
-- name: Age 1 in Oregon, not eligible due to residency.
-  period: 2025
-  input:
-    age: 1
-    state_code: OR
-  output:
-    wa_early_eceap_age_eligible: false
